fix(): use a completely different package to generate blockies
