+++ package.json
@@ -31,7 +31,7 @@
     "autolinker": "^3.14.3",
     "eslint": "^7.32.0",
     "eth-ens-namehash": "^2.0.8",
-    "ethereum-blockies": "^0.1.1",
+    "ethereum-blockies-base64": "^1.0.2",
     "fortmatic": "^2.2.1",
     "graphql": "^15.5.1",
     "graphql-tag": "^2.12.4",
