+++ src/components/Ui/Blockie.vue
@@ -1,0 +1,19 @@
+<script setup>
+import { computed } from 'vue';
+
+import makeBlockie from 'ethereum-blockies-base64';
+const props = defineProps({
+  seed: {
+    type: String,
+    default: 'randomString'
+  }
+});
+
+const blockie = computed(() => {
+  return makeBlockie(props.seed);
+});
+</script>
+
+<template>
+  <img :src="blockie" />
+</template>
